stateful-proptest: Add ability to override proptest default config
This modifies the stateful-proptest API to allow us to pass in a
`ProptestConfig` value, which lets tests override most of the default
proptest config values.

We're not using this capability in either of our vertical test suites
yet, but I'm planning to make use of this in a later commit.

Change-Id: Iddbd219a12745b4ddfe1bfaffe12d7089e649491
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4908
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>