proptest-stateful: Change dependency spec to enable standalone use
Just saying `workspace = true` lets us avoid explicitly specifying a
version when we already have a version in use in the workspace, but
since we're trying to publish proptest-stateful as a standalone library,
this doesn't really cut it anymore. It's possible that versions older
than 1.28 would work fine, but since this is what we're currently using
in the workspace, I'm just sticking with this as a version spec for now
to keep things simple.

Change-Id: I4407237f14de5c2e727903a1c42c62c7033a6862
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4977
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>