proptest-stateful: Add metadata to Cargo.toml for standalone release
These fields are all required or recommended for publishing a crate to
crates.io. Note that the repository and docs page don't actually exist
yet, but these will be created soon, and the URLs should remain the
same.

Also note that we're releasing this crate directly under Apache 2.0,
instead of doing the BSL license and converting to Apache 2.0 like we do
with ReadySet itself. I ran this by Gautam and Alana last night and got
their approval.

Change-Id: Ib366e8ee56cd124f7ae4538b64a95a3e74420bbe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4979
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>