treewide: Upgrade tokio to 1.32
Upgrade the version of tokio we depend on to version 1.32, to get the
version with https://github.com/tokio-rs/tokio/pull/5925, my fix for a
performance issue in `tokio::sync::broadcast`. We use this to notify
workers when channels are removed from the channel coordinator, so we
want this fix to improve the performance of that process.

Change-Id: Id45707e0f95ca0a76ea69952ea23e8c65f846983
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5883
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>