treewide: Upgrade dependencies
This commit upgrades many of our dependencies to newer versions. Note
that this was not just the result of a `cargo update` invocation; I used
the `cargo-edit` tool to automate the process of upgrading crate
versions in our Cargo.toml (I also ran `cargo update` afterwards for
good measure). The code changes in this commit reflect breaking API
changes in the new package versions.

Change-Id: Ib15333b66c6bba2e3eb4a302ea85c3a03ab0acf5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7378
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>